+++ src/components/manage/Delete/Delete.test.jsx
@@ -1,15 +1,19 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import configureStore from 'redux-mock-store';
+import thunk from 'redux-thunk';
 import { Provider } from 'react-intl-redux';
 import { MemoryRouter } from 'react-router-dom';
 
 import Delete from './Delete';
 
-const mockStore = configureStore();
+const mockStore = configureStore([thunk]);
 
 jest.mock('react-portal', () => ({
   Portal: jest.fn(() => <div id="Portal" />),
+}));
+jest.mock('../../../actions/content/content', () => ({
+  getContent: jest.fn(() => ({ type: 'GET_CONTENT' })),
 }));
 
 describe('Delete', () => {
@@ -30,7 +34,10 @@ describe('Delete', () => {
     const component = renderer.create(
       <Provider store={store}>
         <MemoryRouter>
-          <Delete location={{ pathname: '/blog', search: {} }} />
+          <Delete
+            location={{ pathname: '/blog', search: {} }}
+            deleteContent={() => {}}
+          />
         </MemoryRouter>
       </Provider>,
     );
@@ -57,7 +64,10 @@ describe('Delete', () => {
     const component = renderer.create(
       <Provider store={store}>
         <MemoryRouter>
-          <Delete location={{ pathname: '/blog', search: {} }} />
+          <Delete
+            location={{ pathname: '/blog', search: {} }}
+            deleteContent={() => {}}
+          />
         </MemoryRouter>
       </Provider>,
     );